feature: Add content_type field to Toot struct
This commit introduces a new field, ContentType, to the Toot struct in
mastodon.go.

This field is intended to store the MIME type of the content associated
with a Toot, such as "text/plain", "text/html", or "text/markdown".

Creating Markdown posts is supported on Pleroma/Akkoma and some Mastodon
forks.